Fix compatibility with ES 5
ES version 5 comes with a new default regarding mapping of string
fields. The field itself is text and a subfield (keyword) is not
analysed. To perform exact match operations, the keyword subquery needs
to be used.

By doing this, we force the use of ES 5+ and prevent customization of
the mapping, for instance, to remove the intermediary analized field. A
better fix might be to read the mappings and detect keyword subfield
when present.

See https://www.elastic.co/blog/strings-are-dead-long-live-strings